Fix issue 41.
The problem turns out to be caused by a silly bug in the implementation 
of get_{m/p}_structure: in the case that the incoming term is directly 
a free variable (no head normalization is needed), no checking was made 
on its and the structure head's universe counts. The fix is to add the 
check and to fail when the universe count of the structure head is 
greater than that of the free variable.